app/requirements.txt
# Core dependencies
gradio>=4.19.0
devicetorch

# TTS Engines
kokoro>=0.9.2
chatterbox-tts

# PDF Processing
PyMuPDF>=1.23.0
pdf2image>=1.16.0

# OCR
paddlepaddle
paddleocr>=2.7.0

# Image Processing
pillow>=10.0.0
numpy>=1.24.0

# Audio Processing
pydub>=0.25.0
soundfile>=0.12.0
librosa>=0.10.0

# Database
sqlalchemy>=2.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Progress & Logging
tqdm>=4.65.0
rich>=13.0.0
